Extend HEAD response portability coverage
